heroku 14

+++ db/migrate/20160211231347_add_rate_to_progre.rb
@@ -1,5 +1,5 @@
 class AddRateToProgre < ActiveRecord::Migration
   def change
-    add_column :progres, :rate, :double
+    add_column :progres, :rate, :decimal
   end
 end
